Rename: selectPokemon 으로 컴포넌트 이동
