+++ src/components/cardMaker/CardPage.tsx
@@ -1,6 +1,6 @@
 import useSelectedPokemonForCard from '@/store/useSelectedPokemonForCard';
-import CardsRowLike from '@/components/cardMaker/CardsRowLike';
-import CardsRowRandom from '@/components/cardMaker/CardsRowRandom';
+import CardsRowLike from '@/components/cardMaker/selectPokemon/CardsRowLike';
+import CardsRowRandom from '@/components/cardMaker/selectPokemon/CardsRowRandom';
 import styles from './cards.module.scss';
 import PokemonCard from '../card/PokemonCard';
 import CardEditor from './CardEditor';
