+++ src/components/cardMaker/selectPokemon/CardsRowRandom.tsx
